+++ sdk/containerregistry/Azure.Containers.ContainerRegistry/src/ContainerRegistryClient.cs
@@ -2,10 +2,8 @@
 // Licensed under the MIT License.
 
 using System;
-using System.Collections.Specialized;
 using System.Threading;
-using System.Web;
-
+using System.Threading.Tasks;
 using Azure.Core;
 using Azure.Core.Pipeline;
 
@@ -56,69 +54,94 @@ namespace Azure.Containers.ContainerRegistry
         /// <param name="cancellationToken"> The cancellation token to use. </param>
         public virtual AsyncPageable<string> GetRepositoriesAsync(CancellationToken cancellationToken = default)
         {
-            return PageResponseEnumerator.CreateAsyncEnumerable(async (continuationToken, pageSizeHint) =>
+            async Task<Page<string>> FirstPageFunc(int? pageSizeHint)
             {
                 using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(ContainerRegistryClient)}.{nameof(GetRepositories)}");
                 scope.Start();
                 try
                 {
-                    Response<Repositories> response =
-                        await _restClient.GetRepositoriesAsync(
-                            continuationToken,
-                            pageSizeHint,
-                            cancellationToken)
-                       .ConfigureAwait(false);
-
-                    string lastRepository = null;
-                    if (!string.IsNullOrEmpty(response.Value.Link))
-                    {
-                        Uri nextLink = new Uri(response.Value.Link);
-                        NameValueCollection queryParams = HttpUtility.ParseQueryString(nextLink.Query);
-                        lastRepository = queryParams["last"];
-                    }
-
-                    return Page<string>.FromValues(response.Value.RepositoriesValue, lastRepository, response.GetRawResponse());
+                    ResponseWithHeaders<Repositories, ContainerRegistryGetRepositoriesHeaders> response = await _restClient.GetRepositoriesAsync(last: null, n: pageSizeHint, cancellationToken).ConfigureAwait(false);
+                    return Page.FromValues(response.Value.RepositoriesValue, response.Headers.Link, response.GetRawResponse());
                 }
-                catch (Exception ex)
+                catch (Exception e)
                 {
-                    scope.Failed(ex);
+                    scope.Failed(e);
                     throw;
                 }
-            });
+            }
+
+            async Task<Page<string>> NextPageFunc(string continuationToken, int? pageSizeHint)
+            {
+                using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(ContainerRegistryClient)}.{nameof(GetRepositories)}");
+                scope.Start();
+                try
+                {
+                    string uriReference = ParseUriReferenceFromLinkHeader(continuationToken);
+                    ResponseWithHeaders<Repositories, ContainerRegistryGetRepositoriesHeaders> response = await _restClient.GetRepositoriesNextPageAsync(uriReference, last: null, n: pageSizeHint, cancellationToken).ConfigureAwait(false);
+                    return Page.FromValues(response.Value.RepositoriesValue, response.Headers.Link, response.GetRawResponse());
+                }
+                catch (Exception e)
+                {
+                    scope.Failed(e);
+                    throw;
+                }
+            }
+
+            return PageableHelpers.CreateAsyncEnumerable(FirstPageFunc, NextPageFunc);
         }
 
         /// <summary> List repositories. </summary>
         /// <param name="cancellationToken"> The cancellation token to use. </param>
         public virtual Pageable<string> GetRepositories(CancellationToken cancellationToken = default)
         {
-            return PageResponseEnumerator.CreateEnumerable((continuationToken, pageSizeHint) =>
+            Page<string> FirstPageFunc(int? pageSizeHint)
             {
                 using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(ContainerRegistryClient)}.{nameof(GetRepositories)}");
                 scope.Start();
                 try
                 {
-                    Response<Repositories> response =
-                        _restClient.GetRepositories(
-                            continuationToken,
-                            pageSizeHint,
-                            cancellationToken);
-
-                    string lastRepository = null;
-                    if (!string.IsNullOrEmpty(response.Value.Link))
-                    {
-                        Uri nextLink = new Uri(response.Value.Link);
-                        NameValueCollection queryParams = HttpUtility.ParseQueryString(nextLink.Query);
-                        lastRepository = queryParams["last"];
-                    }
-
-                    return Page<string>.FromValues(response.Value.RepositoriesValue, lastRepository, response.GetRawResponse());
+                    ResponseWithHeaders<Repositories, ContainerRegistryGetRepositoriesHeaders> response = _restClient.GetRepositories(last: null, n: pageSizeHint, cancellationToken);
+                    return Page.FromValues(response.Value.RepositoriesValue, response.Headers.Link, response.GetRawResponse());
                 }
-                catch (Exception ex)
+                catch (Exception e)
                 {
-                    scope.Failed(ex);
+                    scope.Failed(e);
                     throw;
                 }
-            });
+            }
+
+            Page<string> NextPageFunc(string continuationToken, int? pageSizeHint)
+            {
+                using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(ContainerRegistryClient)}.{nameof(GetRepositories)}");
+                scope.Start();
+                try
+                {
+                    string uriReference = ParseUriReferenceFromLinkHeader(continuationToken);
+                    ResponseWithHeaders<Repositories, ContainerRegistryGetRepositoriesHeaders> response = _restClient.GetRepositoriesNextPage(uriReference, last: null, n: pageSizeHint, cancellationToken);
+                    return Page.FromValues(response.Value.RepositoriesValue, response.Headers.Link, response.GetRawResponse());
+                }
+                catch (Exception e)
+                {
+                    scope.Failed(e);
+                    throw;
+                }
+            }
+
+            return PageableHelpers.CreateEnumerable(FirstPageFunc, NextPageFunc);
+        }
+
+        private static string ParseUriReferenceFromLinkHeader(string linkValue)
+        {
+            // Per the Docker v2 HTTP API spec, the Link header is an RFC5988
+            // compliant rel='next' with URL to next result set, if available.
+            // See: https://docs.docker.com/registry/spec/api/
+            //
+            // The URI reference can be obtained from link-value as follows:
+            //   Link       = "Link" ":" #link-value
+            //   link-value = "<" URI-Reference ">" * (";" link-param )
+            // See: https://tools.ietf.org/html/rfc5988#section-5
+
+            return linkValue?.Substring(1, linkValue.IndexOf('>') - 1);
         }
     }
 }