+++ sdk/storage/Azure.Storage.Files.DataLake/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 12.7.0-beta.3 (Unreleased)
 - Added support for service version 2020-08-04.
 - Added support for Soft Delete for Hierarchical-Namespace enabled accounts.
+- Added support for File Query Parquet input format.
 - Aligned storage URL parsing with other platforms.
 
 ## 12.7.0-beta.2 (2021-03-09)